+++ changelog.md
@@ -1,5 +1,6 @@
 ##### Version 1.21.1-12.4.2-pre
 - Fix mixer accidentally crafting a lingering awkward potion (BluSunrize)
+- Fix catwalks not working properly when using a hammer (BluSunrize)
 
 ##### Version 1.21.1-12.4.1-193
 - Add variable lightlevel for lanterns and cage lamps (BluSunrize)